setup react components
